feature_ingestion/app/requirements.txt
pandas
sqlalchemy
openpyxl
psycopg2-binary
python-dotenv
pymysql